MAINT Merge math_util into util module
I no longer need a separate module for utility functions and
mathematical functions. This is an anachronism from when I was using
Cython for the optimisation of some functions.
